requirements.txt
google-genai>=1.0.0
pydantic>=2.0
openai-whisper
sounddevice
scipy
numpy
pyttsx3
python-dotenv
rich
requests
